chore: consolidate `useScript()` back into core, drop legacy support
