--- docs/scripts/0.introduction.md
@@ -1,97 +1,0 @@
----
-title: "Loading Scripts with Unhead"
-description: "Add type-safe scripts to your site - optimized for SSR, privacy and performance"
-navigation:
-  title: Introduction
----
-
-# Loading Scripts with Unhead
-
-Third-party scripts slow down your site and leak user data. Unhead Scripts helps you fix this.
-
-## Install
-
-```bash
-pnpm add -D @unhead/scripts
-```
-
-## Quick Example
-
-```ts
-const script = useScript('https://example.com/tracker.js', {
-  use() {
-    return window.tracker
-  }
-})
-
-// Works in SSR, runs when script loads
-script.proxy.trackEvent('page_view')
-```
-
-## What Makes It Different?
-
-- Scripts don't block page load
-- No script runs until you need it
-- No user data sent without consent
-- Works in SSR without hacks
-
-## When Not to Use
-
-1. **Static Scripts**: If your script never changes and runs on every page, just add it to your HTML.
-2. **Small Utilities**: For scripts under 5KB, just bundle them.
-3. **Core Features**: Don't lazy load scripts your site needs to work.
-
-## Real Performance Impact
-
-Loading scripts the wrong way makes your site slow:
-
-- 1 analytics script: +2s load time
-- 1 chat widget: +3s load time
-- 1 ad script: +4s load time
-
-Source: [HTTP Archive's Web Almanac](https://almanac.httparchive.org/en/2023/performance)
-
-## External Resources
-
-- [Script Loading Best Practices (web.dev)](https://web.dev/optimizing-content-efficiency-loading-third-party-javascript/)
-- [Performance Impact of Third Party JS](https://www.speedcurve.com/blog/real-cost-of-third-party-javascript/)
-- [Browser Resource Loading (MDN)](https://developer.mozilla.org/en-US/docs/Web/Performance/Resource_loading)
-
-## Next Steps
-
-- [Add Types to Scripts →](/unhead/scripts/typed-scripts)
-- [Control Loading →](/unhead/scripts/load-triggers)
-- [Handle Errors →](/unhead/scripts/load-failures)
-
----
-
-title: Introduction to Unhead
-description: Manage page metadata for modern JavaScript applications
-navigation:
-title: Introduction
----
-
-## What is head manager?
-
-Unhead is built for modern JavaScript applications that need to render code outside their `<div id="app"></div>`{lang="html"} entry in both a
-server-rendered and client-rendered environment.
-
-JavaScript applications need different metadata for each page or view. Without proper management, applications encounter several issues:
-
-Pages show incorrect titles and descriptions during navigation. Social media previews display wrong information. Search engines receive duplicate or conflicting meta tags. Server-rendered pages break when JavaScript takes over in the browser.
-
-These issues affect SEO, social sharing, and user experience. They become more significant in server-rendered applications where metadata must work correctly during the initial page load and subsequent JavaScript updates.
-
-```html
-<!DOCTYPE html>
-<html>  <!-- ? htmlAttrs -->
-  <head>
-    <!-- ? head -->
-  </head>
-  <body> <!-- Unhead: bodyAttrs -->
-    <!-- Unhead: bodyOpen -->
-    <div id="app"></div> <!-- Your app -->
-    <!-- Unhead: bodyClose -->
-  </body>
-</html>
-```
