--- docs/scripts/0.nuxt/.navigation.yml
@@ -1,2 +1,0 @@
-title: Vue
-icon: logos:vue
