--- docs/scripts/0.typescript/1.installation.md
@@ -1,191 +1,0 @@
----
-title: 'Install Unhead'
-description: 'Get started with Unhead by installing the dependency to your project.'
-navigation:
-  title: 'Installation'
----
-
-Using :Icon{name="logos:vue"} Vue? Check out the [Vue integration](/setup/vue/installation).
-
-Using :Icon{name="logos:nuxt-icon"} Nuxt? Unhead is already built-in! Check out the [Nuxt docs](https://nuxt.com/docs/getting-started/seo-meta).
-
-## Setup
-
-1. Install `unhead` dependency to your project:
-
-::code-group
-
-```bash [yarn]
-yarn add unhead
-```
-
-```bash [npm]
-npm install unhead
-```
-
-```bash [pnpm]
-pnpm add unhead
-```
-
-::
-
-### Setup Client-Side Rendering (CSR)
-
-Create a head instance somewhere in your apps client entry.
-
-```ts [main.ts]
-import { createHead } from 'unhead/client'
-
-// Create a global head instance
-const head = createHead()
-```
-
-Done! Now you can use the `useHead` composable to manage your head.
-
-```ts
-import { useHead } from 'unhead'
-
-useHead({
-  title: 'My awesome site'
-})
-```
-
-### Setup Server-Side Rendering (SSR)
-
-If you app does not SSR, you can skip this step.
-
-**1. Create a head instance**
-
-Somewhere in your server entry, create a head instance.
-
-```ts [main.ts]
-import { createHead } from 'unhead/server'
-
-// Create a global head instance
-const head = createHead()
-```
-
-**2. Manage context**
-
-To make use of the `useHead()`{lang="ts"} composables, Unhead uses [unctx](https://github.com/unjs/unctx) under the hood which requires you to manually
-set and unset the context when handling requests to avoid cross-request pollution.
-
-This is usually done by attaching setting the context as part of handling the request.
-
-```ts
-import { unheadCtx } from 'unhead'
-import { createHead } from 'unhead/server'
-
-function handleRequest(req, res) {
-  const head = createHead()
-  unheadCtx.set(head)
-  // Your SSR logic here
-  unheadCtx.unset()
-}
-```
-
-**3. Generate SSR Payload**
-
-Once you're ready to start displaying tags on the server, you'll need to generate the SSR payload.
-
-```ts
-import { renderSSRHead } from 'unhead/server'
-
-// head is from createHead()
-const payload = await renderSSRHead(head)
-```
-
-The payload schema looks like the following:
-
-```ts
-export interface SSRHeadPayload {
-  headTags: string
-  bodyTags: string
-  bodyTagsOpen: string
-  htmlAttrs: string
-  bodyAttrs: string
-}
-```
-
-**4. Update your app template**
-
-You will need to update your app template to add in the templates for
-the SSR tags.
-
-Different frameworks differ in how they handle this template.
-
-**Simple string replace**
-
-```html
-<!DOCTYPE html>
-<html<!--htmlAttrs-->>
-  <head>
-    <!--headTags-->
-    <!--preload-links-->
-  </head>
-  <body<!--bodyAttrs-->>
-    <!--bodyTagsOpen-->
-    <div id="app"><!--app-html--></div>
-    <script type="module" src="/src/entry-client.js"></script>
-    <!--bodyTags-->
-  </body>
-</html>
-```
-
-To handle this type of template you can use this code
-
-```ts
-const headPayload = await renderSSRHead(head)
-
-Object.entries(headPayload).forEach(([key, value]) => {
-  html = html.replace(`<!--${key}-->`{lang="html"}, value)
-})
-```
-
-## 3. Done! How hydration works
-
-When your client-side app hydrates the server head tags, it will attempt to hydrate each
-element based on the nodes being equal `$el.isEqualNode($newEl)` or them sharing the same
-dedupe key (see [Tag Deduping](/usage/guides/handling-duplicates)).
-
-If you're rendering content that differs between client and server, you should
-specify a `key` attribute if the element does not have a unique dedupe key.
-
-```ts
-useHead({
-  script: [
-    {
-      // key is needed to avoid seperate scripts being created
-      key: 'my-script',
-      innerHTML: process.server ? '' : 'console.log("hello world")',
-    }
-  ]
-})
-```
-
-## Optional: Auto-Imports
-
-Unhead provides out-of-the-box configuration for [unplugin-auto-import](https://github.com/antfu/unplugin-auto-import).
-
-```ts [vite.config.ts]
-import { unheadComposablesImports } from 'unhead'
-
-export default defineConfig({
-  plugins: [
-    AutoImport({
-      imports: [
-        unheadComposablesImports[0],
-      ],
-    }),
-  ]
-})
-```
-
-### Next Steps
-
-Your app is now setup for head management, congrats! 🎉
-
-Try next:
-
-1. Optional: [Setup SSR](/setup/ssr/installation)
-2. Add some [recipes](/addons/recipes)
