--- docs/scripts/1.guides/.navigation.yml
@@ -1,2 +1,0 @@
-title: Guides
-icon: noto:open-book
