--- docs/scripts/1.guides/3.consent.md
@@ -1,80 +1,0 @@
----
-title: Consent Management
-description: Learn how to get user consent before loading scripts.
----
-
-## Background
-
-Many third-party scripts include tracking cookies that require user consent under privacy laws. Nuxt Scripts simplifies this process with the [useScriptTriggerConsent](/docs/api/use-script-trigger-consent) composable, allowing scripts to be loaded only after receiving user consent.
-
-## Usage
-
-The [useScriptTriggerConsent](/docs/api/use-script-trigger-consent) composable offers flexible interaction options suitable for various scenarios.
-
-See the [API](/docs/api/use-script-trigger-consent) docs for full details on the available options.
-
-### Accepting as a Function
-
-The easiest way to make use of `useScriptTriggerConsent` is by invoking the `accept` method when user consent is granted.
-
-For an example of how you might lay out your code to handle this, see the following:
-
-::code-group
-
-```ts [utils/cookie.ts]
-export const agreedToCookiesScriptConsent = useScriptTriggerConsent()
-```
-
-```vue [app.vue]
-<script setup lang="ts">
-import { agreedToCookiesScriptConsent } from '#imports'
-
-useScript('https://www.google-analytics.com/analytics.js', {
-  trigger: agreedToCookiesScriptConsent
-})
-</script>
-```
-
-```vue [components/cookie-banner.vue]
-<script setup lang="ts">
-import { agreedToCookiesScriptConsent } from '#imports'
-</script>
-
-<template>
-  <button @click="agreedToCookiesScriptConsent.accept()">
-    Accept Cookies
-  </button>
-</template>
-```
-
-::
-
-### Accepting as a resolvable boolean
-
-Alternatively, you can pass a reactive reference to the consent state to the `useScriptTriggerConsent` composable. This will automatically load the script when the consent state is `true`.
-
-```ts
-const agreedToCookies = ref(false)
-useScript('https://www.google-analytics.com/analytics.js', {
-  trigger: useScriptTriggerConsent({
-    consent: agreedToCookies
-  })
-})
-```
-
-### Delaying script load after consent
-
-There may be instances where you want to trigger the script load after a certain event, only if the user has consented.
-
-For this you can use the `postConsentTrigger`, which shares the same API as `trigger` from the `useScript` composable.
-
-```ts
-const agreedToCookies = ref(false)
-useScript('https://www.google-analytics.com/analytics.js', {
-  trigger: useScriptTriggerConsent({
-    consent: agreedToCookies,
-    // load 3 seconds after consent is granted
-    postConsentTrigger: new Promise<void>(resolve => setTimeout(resolve, 3000))
-  })
-})
-```
