--- packages/scripts/README.md
@@ -1,7 +1,0 @@
-# @unhead/scripts
-
-Unhead Scripts allows you to load third-party scripts with better performance, privacy, and security.
-
-## License
-
-MIT License © 2022-PRESENT [Harlan Wilton](https://github.com/harlan-zw)
