--- packages/scripts/build.config.ts
@@ -1,11 +1,0 @@
-import { defineBuildConfig } from 'unbuild'
-
-export default defineBuildConfig({
-  clean: true,
-  declaration: true,
-  entries: [
-    { input: 'src/index' },
-    { input: 'src/vue/index', name: 'vue' },
-  ],
-  externals: ['vue'],
-})
