--- packages/scripts/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "@unhead/scripts",
-  "type": "module",
-  "version": "2.0.0-alpha.16",
-  "description": "Unhead Scripts allows you to load third-party scripts with better performance, privacy, and security.",
-  "author": "Harlan Wilton <harlan@harlanzw.com>",
-  "license": "MIT",
-  "funding": "https://github.com/sponsors/harlan-zw",
-  "publishConfig": {
-    "access": "public",
-    "tag": "next"
-  },
-  "homepage": "https://unhead.unjs.io",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/unjs/unhead.git",
-    "directory": "packages/scripts"
-  },
-  "bugs": {
-    "url": "https://github.com/unjs/unhead/issues"
-  },
-  "sideEffects": false,
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs"
-    },
-    "./vue": {
-      "types": "./dist/vue.d.ts",
-      "import": "./dist/vue.mjs"
-    }
-  },
-  "main": "dist/index.mjs",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "typesVersions": {
-    "*": {
-      "vue": [
-        "dist/vue"
-      ]
-    }
-  },
-  "files": [
-    "dist",
-    "overrides.d.ts",
-    "vue.d.ts"
-  ],
-  "scripts": {
-    "build": "unbuild .",
-    "stub": "unbuild . --stub",
-    "test": "vitest",
-    "release": "bumpp package.json --commit --push --tag",
-    "lint": "eslint \"{src,test}/**/*.{ts,vue,json,yml}\" --fix"
-  },
-  "peerDependencies": {
-    "@unhead/vue": "workspace:*",
-    "unhead": "workspace:*"
-  },
-  "peerDependenciesMeta": {
-    "@unhead/vue": {
-      "optional": true
-    },
-    "unhead": {
-      "optional": true
-    }
-  }
-}
