+++ packages/unhead/build.config.ts
@@ -13,6 +13,7 @@ export default defineBuildConfig({
     { input: 'src/plugins/index', name: 'plugins' },
     { input: 'src/server/index', name: 'server' },
     { input: 'src/client/index', name: 'client' },
+    { input: 'src/scripts/index', name: 'scripts' },
     { input: 'src/legacy', name: 'legacy' },
     { input: 'src/types', name: 'types' },
     { input: 'src/utils', name: 'utils' },
