+++ packages/unhead/src/scripts/index.ts
