+++ packages/unhead/src/scripts/types.ts
