+++ packages/unhead/src/scripts/utils.ts
