+++ packages/unhead/src/types/index.ts
@@ -1,7 +1,8 @@
-export * from './head'
-export * from './hooks'
-export * from './plugins'
-export * from './safeSchema'
-export * from './schema'
-export * from './tags'
-export * from './util'
+export type * from '../scripts'
+export type * from './head'
+export type * from './hooks'
+export type * from './plugins'
+export type * from './safeSchema'
+export type * from './schema'
+export type * from './tags'
+export type * from './util'
