+++ packages/unhead/test/unit/scripts/warmup.test.ts
@@ -1,7 +1,7 @@
 import type { LinkBase } from 'unhead/types'
 import { createHead as createServerHead } from 'unhead/server'
 import { describe, it } from 'vitest'
-import { useScript } from '../../src/useScript'
+import { useScript } from '../../../src/composables'
 
 describe('warmup', () => {
   it('server', () => {
