+++ packages/vue/src/scripts/index.ts
