+++ packages/vue/src/scripts/useScript.ts
@@ -1,4 +1,5 @@
 import type { ResolvableProperties, VueHeadClient } from '@unhead/vue'
+import type { UseScriptOptions as BaseUseScriptOptions, ScriptInstance, UseFunctionType, UseScriptStatus } from 'unhead/scripts'
 import type {
   DataKeys,
   HeadEntryOptions,
@@ -6,10 +7,9 @@ import type {
   ScriptBase,
 } from 'unhead/types'
 import type { ComponentInternalInstance, Ref, WatchHandle } from 'vue'
-import type { UseScriptOptions as BaseUseScriptOptions, ScriptInstance, UseFunctionType, UseScriptStatus } from '../types'
 import { injectHead } from '@unhead/vue'
+import { useScript as _useScript } from 'unhead/scripts'
 import { getCurrentInstance, isRef, onMounted, onScopeDispose, ref, watch } from 'vue'
-import { useScript as _useScript } from '../useScript'
 
 export interface VueScriptInstance<T extends Record<symbol | string, any>> extends Omit<ScriptInstance<T>, 'status'> {
   status: Ref<UseScriptStatus>
